Improvements to Western language char rendering
- Fixed problems with rendering on high res screens
- Use UTF8 encoding for all source files
- Rename WesternEuropean to WesternCharSet

+++ Marlin/src/lcd/extensible_ui/lib/lulzbot/ftdi_eve_lib/extended/unicode/standard_char_set.h
@@ -23,7 +23,7 @@ namespace FTDI {
   class StandardCharSet {
     public:
       static uint8_t std_char_width(char);
-      static void load_data(uint16_t addr);
+      static void load_data(uint32_t addr);
       static void load_bitmaps(CommandProcessor&);
       static bool render_glyph(CommandProcessor*, int &x, int &y, font_size_t, utf8_char_t);
   };
